Fix: 벙 수정시 주최자 변경 및 목록에서 삭제할 경우 에러 수정
- service 딴에서 FK 이슈로 인하여 transaction 순서 변경

+++ backend/src/main/java/com/admin/the_climbing_night/meeting/service/MeetingService.java
@@ -78,12 +78,19 @@ public class MeetingService {
 
     @Transactional
     public int updateMeeting(List<String> deleteAttends, List<InsertAttendVo> insertAttendsVo,
-            UpdateMeetingRequest req) {
+            UpdateMeetingRequest req) throws Exception {
+
+        int result = meetingMapper.updateMeeting(req);
+
+        if (result < 1) {
+            throw new Exception("Meeting Update Fail");
+        }
+
         if (deleteAttends.size() > 0)
             meetingMapper.deleteAttends(deleteAttends);
         if (insertAttendsVo.size() > 0)
             meetingMapper.insertAttends(insertAttendsVo);
 
-        return meetingMapper.updateMeeting(req);
+        return result;
     }
 }
